+++ packages/arui-scripts/src/commands/util/exec.ts
@@ -2,6 +2,7 @@ import shell from 'shelljs';
 
 function exec(command: string) {
     return new Promise((resolve, reject) => {
+        console.log(`Executing command: ${command}`);
         shell.exec(command, (code) => {
             if (code === 0) {
                 return resolve(code);