+++ packages/arui-scripts/src/index.ts
@@ -1,3 +1,5 @@
 export type { OverrideFile } from './configs/util/apply-overrides';
 
 export type { PackageSettings } from './configs/app-configs/available-package-settings';
+export { prepareFilesForDocker } from './commands/util/docker-build';
+export { getBuildParamsFromArgs } from './commands/util/docker-build';