+++ packages/arui-scripts/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [13.3.0-feat-fast-docker-build.3](https://git.moscow.alfaintra.net/ef/arui-scripts/compare/v13.3.0-feat-fast-docker-build.2...v13.3.0-feat-fast-docker-build.3) (2022-12-06)
+
+
+### Bug Fixes
+
+* **docker-build-fast:** don't remove build directory ([0909182](https://git.moscow.alfaintra.net/ef/arui-scripts/commit/09091820f10fffc54702ae6572f6c9a716fe0784))
+
+# [13.3.0-feat-fast-docker-build.2](https://git.moscow.alfaintra.net/ef/arui-scripts/compare/v13.3.0-feat-fast-docker-build.1...v13.3.0-feat-fast-docker-build.2) (2022-12-06)
+
+
+### Features
+
+* **docker-fast:** don't build anything inside of container ([8f255bd](https://git.moscow.alfaintra.net/ef/arui-scripts/commit/8f255bd8fb9818a8498bddee71ef83ce4a7ba11b))
+
+# [13.3.0-feat-fast-docker-build.1](https://git.moscow.alfaintra.net/ef/arui-scripts/compare/v13.2.1...v13.3.0-feat-fast-docker-build.1) (2022-12-05)
+
+
+### Features
+
+* **docker:** add fast docker build command ([82301a8](https://git.moscow.alfaintra.net/ef/arui-scripts/commit/82301a8d912476fd5516f94083ccd7c5363f4509))
+
 ## [13.2.1](https://git.moscow.alfaintra.net/ef/arui-scripts/compare/v13.2.0...v13.2.1) (2022-12-05)